fachschaftszitat/requirements.txt
Django==2.0.1
django-jinja==2.4.1
Jinja2==2.10
MarkupSafe==1.0
pytz==2017.3
djangorestframework==3.9.0
